Added Port Number Flag
Added port flag when running command so we can just pass of -p flag to specify the authrpc port to run the listen to when a node goes up. Set defaults as node1 and 8546 node and port name respectively. Can change later on.

Eg run the following command in the ec2 instance: 

bash node_setup.sh -n nodeX -p 8548